+++ src/trapz/trapz.h
@@ -4,6 +4,6 @@
 #include <math.h>
 #include <stdint.h>
 
-double catools_trapz(uint16_t n, double *x, double *y);
+double catools_trapz(uint32_t n, double *x, double *y);
 
 #endif // _TRAPZ_H_
